Refactor validation modules (#542)
* Modify call to validation module so it uses the task manager and leverage the cache

* Reduce batch size in validation module

* Rename validation folder

* Wait for validation to be done before continuing
